django_chat/requirements.txt
django == 4.2.3
mysqlclient == 2.2.0
djangorestframework == 3.14.0